addon/requirements.txt
requests
pydantic
retry
paho-mqtt
pytcs-tecnoalarm
